Add some notes about compiler/linter errors
This is not intended to be merged as is, but to start the discussion around these things. In fact, it isn't clear that this is necessarily "style" at all so this may not be an appropriate place for these. On the other hand, it _might_ be useful to document this somewhere.

Some notes:
- The language around compiler errors (warnings too, but `-Werror`) is purposefully more strict than the lint errors. This is to convey that, especially in a "the world is on fire"  type situation, the linter errors are less important than compiler ones.
- Also mentioned, so it is hopefully remembered, when we can finally get to ghc >= 8.4.1 to change our `-Wall` to `-Weverything` since `all != everything` :/